split velodrome addresses into addresses and config
